+++ src/Simplify.Templates.Tests/TemplateBuilderLocalizationTests.cs
@@ -12,15 +12,15 @@ public class TemplateBuilderLocalizationTests
 	private const string LocalTestFilePath = "TestTemplates/Local/LocalizationTest.tpl";
 	private const string EmbeddedTestFilePath = "TestTemplates/Embedded/LocalizationTest.tpl";
 
-	private static readonly object[] _testCases =
-	{
+	private static readonly object[] TestCases =
+	[
 		(TemplateBuilderDelegate) (() => TemplateBuilder.FromFile(FileUtil.ConstructFullFilePath(LocalTestFilePath))),
 		(TemplateBuilderDelegate) (() => TemplateBuilder.FromLocalFile(LocalTestFilePath)),
 		(TemplateBuilderDelegate) (() => TemplateBuilder.FromAssembly(EmbeddedTestFilePath, Assembly.GetExecutingAssembly())),
 		(TemplateBuilderDelegate) (() => TemplateBuilder.FromCurrentAssembly(EmbeddedTestFilePath))
-	};
+	];
 
-	[TestCaseSource(nameof(_testCases))]
+	[TestCaseSource(nameof(TestCases))]
 	public void Build_LocalizableDifferentFromBase_LocalizableLoadedWithBaseReplacements(TemplateBuilderDelegate templateBuilder)
 	{
 		// Arrange
@@ -32,10 +32,10 @@ public class TemplateBuilderLocalizationTests
 			.Build();
 
 		// Assert
-		Assert.AreEqual("текст1 text2", tpl.Get());
+		Assert.That(tpl.Get(), Is.EqualTo("текст1 text2"));
 	}
 
-	[TestCaseSource(nameof(_testCases))]
+	[TestCaseSource(nameof(TestCases))]
 	public async Task BuildAsync_LocalizableDifferentFromBase_LocalizableLoadedWithBaseReplacements(TemplateBuilderDelegate templateBuilder)
 	{
 		// Arrange
@@ -47,10 +47,10 @@ public class TemplateBuilderLocalizationTests
 			.BuildAsync();
 
 		// Assert
-		Assert.AreEqual("текст1 text2", tpl.Get());
+		Assert.That(tpl.Get(), Is.EqualTo("текст1 text2"));
 	}
 
-	[TestCaseSource(nameof(_testCases))]
+	[TestCaseSource(nameof(TestCases))]
 	public void Build_LocalizableFromCurrentThreadLanguageDifferentFromBase_LocalizableLoadedWithBaseReplacements(
 		TemplateBuilderDelegate templateBuilder)
 	{
@@ -65,10 +65,10 @@ public class TemplateBuilderLocalizationTests
 			.Build();
 
 		// Assert
-		Assert.AreEqual("текст1 text2", tpl.Get());
+		Assert.That(tpl.Get(), Is.EqualTo("текст1 text2"));
 	}
 
-	[TestCaseSource(nameof(_testCases))]
+	[TestCaseSource(nameof(TestCases))]
 	public async Task BuildAsync_LocalizableFromCurrentThreadLanguageDifferentFromBase_LocalizableLoadedWithBaseReplacements(
 		TemplateBuilderDelegate templateBuilder)
 	{
@@ -83,6 +83,6 @@ public class TemplateBuilderLocalizationTests
 			.BuildAsync();
 
 		// Assert
-		Assert.AreEqual("текст1 text2", tpl.Get());
+		Assert.That(tpl.Get(), Is.EqualTo("текст1 text2"));
 	}
 }