+++ src/Simplify.Templates/ITemplate.cs
@@ -1,43 +1,41 @@
 ﻿namespace Simplify.Templates;
 
 /// <summary>
-/// Text templates interface
+/// Represents a text template
 /// </summary>
 public interface ITemplate
 {
 	/// <summary>
-	/// Set template variable value (all occurrences will be replaced)
+	/// Sets the  template variable value (all occurrences will be replaced)
 	/// </summary>
-	/// <param name="variableName">Variable name</param>
-	/// <param name="value">Value to set</param>
+	/// <param name="variableName">The variable name</param>
+	/// <param name="value">The value to set</param>
 	ITemplate Set(string variableName, string? value);
 
 	/// <summary>
-	/// Add value to set template variable value (all occurrences will be replaced on Get method execute) allows setting multiple values to template variable
+	/// Adds the value to the list to set to the template variable value (all occurrences will be replaced on Get method execute) allows setting multiple values to template variable
 	/// </summary>
 	/// <param name="variableName">Variable name</param>
 	/// <param name="value">Value to set</param>
 	ITemplate Add(string variableName, string? value);
 
 	/// <summary>
-	/// Get text of the template
+	/// Gets the text of the template
 	/// </summary>
 	string Get();
 
 	/// <summary>
-	/// Sets initial template state equal to current.
+	/// Sets the initial template state equal to current.
 	/// </summary>
-	/// <returns></returns>
 	void Commit();
 
 	/// <summary>
-	/// Return template to it's initial state
+	/// Returns the template to it's initial state
 	/// </summary>
 	void RollBack();
 
 	/// <summary>
 	/// Gets the text of the template and returns loaded template to it's initial state
 	/// </summary>
-	/// <returns>Text of the template</returns>
 	string GetAndRoll();
 }