+++ src/Simplify.Templates/Model/ModelSetterBase.cs
@@ -1,21 +1,15 @@
 ﻿namespace Simplify.Templates.Model;
 
 /// <summary>
-/// Provides ModelSetter base class
+/// Provides the model setter base class
 /// </summary>
-public abstract class ModelSetterBase
+/// <remarks>
+/// Initializes a new instance of the <see cref="ModelSetterBase" /> class.
+/// </remarks>
+/// <param name="template">The template.</param>
+/// <param name="modelPrefix">The model prefix.</param>
+public abstract class ModelSetterBase(ITemplate template, string? modelPrefix = null)
 {
-	/// <summary>
-	/// Initializes a new instance of the <see cref="ModelSetterBase" /> class.
-	/// </summary>
-	/// <param name="template">The template.</param>
-	/// <param name="modelPrefix">The model prefix.</param>
-	protected ModelSetterBase(ITemplate template, string? modelPrefix = null)
-	{
-		ModelPrefix = modelPrefix;
-		Template = template;
-	}
-
 	/// <summary>
 	/// The model prefix separator
 	/// </summary>
@@ -24,20 +18,16 @@ public abstract class ModelSetterBase
 	/// <summary>
 	/// Gets the template.
 	/// </summary>
-	/// <value>
-	/// The template.
-	/// </value>
-	public ITemplate Template { get; }
+	public ITemplate Template { get; } = template;
 
 	/// <summary>
-	/// The model prefix
+	/// Gets the model prefix
 	/// </summary>
-	protected string? ModelPrefix { get; }
+	protected string? ModelPrefix { get; } = modelPrefix;
 
 	/// <summary>
 	/// Formats the name of the variable to replace respecting model prefix.
 	/// </summary>
 	/// <param name="variableName">Name of the variable.</param>
-	/// <returns></returns>
 	protected string FormatModelVariableName(string variableName) => ModelPrefix != null ? ModelPrefix + ModelPrefixSeparator + variableName : variableName;
 }