+++ src/Simplify.Templates/Model/TemplateModelExtensions.cs
@@ -1,7 +1,7 @@
 ﻿namespace Simplify.Templates.Model;
 
 /// <summary>
-/// Template model extensions
+/// Provides the template model extensions
 /// </summary>
 public static class TemplateModelExtensions
 {
@@ -12,7 +12,6 @@ public static class TemplateModelExtensions
 	/// <param name="template">The template.</param>
 	/// <param name="model">The model.</param>
 	/// <param name="modelPrefix">The model prefix (for example 'Model', it will be used like 'Model.YourVariableName').</param>
-	/// <returns></returns>
 	public static IModelSetter<T> Model<T>(this ITemplate template, T model, string? modelPrefix = null)
 		where T : class =>
 		new ModelSetter<T>(template, model, modelPrefix);